removed key-gen and added random generator to hacspec-dev/prelude.rs. keygen functionality is part of test_ntru now
