+++ Cargo.toml
@@ -8,5 +8,4 @@ members = [
     "abstract-integers",
     "spec-examples",
     "attributes",
-    "key-gen",
 ]
