+++ hacspec-dev/Cargo.toml
@@ -7,5 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+rand = "0.7"
+num = "0.2"
 serde_json = "1.0"
 serde = {version = "1.0", features = ["derive"]}
