--- key-gen/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "key_gen"
-version = "0.1.0"
-authors = ["TomerHawk <t.yavorgmx.net>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-rand = "0.7"
-num = "0.2"
-# TODO: serde is only necessary for the test crate/feature
-serde_json = "1.0"
-serde = {version = "1.0", features = ["derive"]}
-
-# TODO: this need to change for releases HERE MAKE CHANGES
-secret_integers = { path = "../secret-integers" }
-abstract_integers = { path = "../abstract-integers" }
-hacspec = {path = "../hacspec/"}
