+++ spec-examples/tests/test_ntru.rs
@@ -1,27 +1,75 @@
 use hacspec::prelude::*;
-
 use hacspec_examples::ntru_prime::*;
-use key_gen::*;
 
-#[test]
-fn test_create_inv() {
-    let mut poly = create_invertable_poly(&ntru_v!(1), ntru_v!(1).q);
-    let mut f_inv: Seq<i128> = Seq::new(ntru_v!(1).p + 1);
+
+
+/// Random byte array
+fn random_byte() -> usize {
+    let random_byte = rand::thread_rng().gen::<usize>();
+    random_byte
+}
+///This function creates a random polynom with w many -1 or 1 and with the highes degree of h_deg.
+fn create_rand_poly(w: usize, h_deg: usize) -> Seq<i128> {
+    let mut counter: usize = 0;
+    let mut pos;
+    let mut polynom: Seq<i128> = Seq::new(h_deg + 1);
+
+    for _ in 0..h_deg * h_deg * h_deg * h_deg *h_deg *h_deg {
+        pos = random_byte() % h_deg;
+        let c_val = (random_byte() % 2) as i128;
+        //check if value already contained
+        if polynom[pos] == 0 {
+            polynom[pos] =  2 * c_val - 1;
+        } else {
+            continue;
+        }
+
+        counter = counter + 1;
+
+        if counter == w as usize {
+            break;
+        }
+    }
+    polynom
+}
+
+fn create_invertable_poly(
+    n: &NtruVersion,
+    modulus: i128,
+) -> (Seq<i128>, Result<Seq<i128>, &'static str>) {
+    let f: Seq<i128> = create_rand_poly(n.w, n.p);
+    let x = eea(&f, &n.irr, modulus);
+    (f, x)
+}
+
+fn key_gen(n_v: &NtruVersion) -> (Seq<i128>, (Seq<i128>, Seq<i128>)) {
+    let mut poly_g = create_invertable_poly(n_v, 3);
+    // Nearly a probability of 1 to find an invertable polynom
+    let mut g_inv: Seq<i128> = Seq::new(n_v.p);
     for _ in 0..4 {
-        match poly.1 {
+        match poly_g.1 {
             Ok(v) => {
-                f_inv = v;
+                g_inv = v;
                 break;
             }
-            Err(_) => poly = create_invertable_poly(&ntru_v!(1), ntru_v!(1).q),
+            Err(_) => poly_g = create_invertable_poly(n_v, 3),
         }
     }
 
-    let result = mul_poly_irr(&f_inv,&poly.0,&(ntru_v!(1).irr),ntru_v!(1).q);
-    assert_eq!(deg(&result),0);
-    if result[0] != 1 as i128{
-        panic!("Didn't work!");
+    //create f
+    let f = create_rand_poly(n_v.w, n_v.p);
+    let f_3times = add_poly(&f, &add_poly(&f, &f, n_v.q,false), n_v.q,false);
+    let f_3times_pre_inv = eea(&f_3times,&n_v.irr,n_v.q);
+    let mut f_inv_3times:Seq<i128> = Seq::new(n_v.p + 1);
+    match f_3times_pre_inv {
+        Ok(v) => {
+            f_inv_3times = v;
+        }
+        Err(_) => println!("Key generating, failed")
     }
+    let h = mul_poly_irr(&poly_g.0, &f_inv_3times,&n_v.irr,n_v.q);
+
+    (h, (f, g_inv))
 }
 
 #[test]
@@ -35,7 +83,12 @@ fn test_round(){
 }
 #[test]
 fn test_encryption_decryption() {
-    let n_v = ntru_v!(0);
+    let n_v = NtruVersion {
+        p: 761,
+        q: 4591,
+        w: 286,
+        irr: set_irr(761),
+    };
     let keys = key_gen(&n_v);
     let pk = keys.0;
     let sk = keys.1;
